refactor: changed the names of several component files
