+++ src/components/BannerSection.vue
