+++ src/components/BenefitSection.vue
