+++ src/components/FaqSection.vue
