+++ src/components/FooterSection.vue
