+++ src/components/PricingSection.vue
