+++ src/components/loggedin/LeftBar.vue
@@ -1,4 +1,4 @@
-<script setup>
+<script setup lang="ts">
 import { ref } from 'vue'
 import { RouterLink } from 'vue-router'
 import { usePocketbaseStore } from '@/stores/pocketbase'
