+++ src/views/HomeView.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
-import Header from '../components/Header.vue'
-import Banner from '../components/Banner.vue'
-import Benefit from '../components/Benefit.vue'
-import Pricing from '../components/Pricing.vue'
-import FAQ from '../components/Faq.vue'
-import Footer from '../components/Footer.vue'
+import Header from '../components/HeaderSection.vue'
+import Banner from '../components/BannerSection.vue'
+import Benefit from '../components/BenefitSection.vue'
+import Pricing from '../components/PricingSection.vue'
+import FAQ from '../components/FaqSection.vue'
+import Footer from '../components/FooterSection.vue'
 </script>
 
 <template>
